Mock loadAndCleanDatabase properly, update tests, and pass strings into GoFSH function.
