+++ src/tests/utils/RunSUSHI.test.js
@@ -3,9 +3,7 @@ import * as processing from '../../utils/Processing';
 import Patient from './fixtures/StructureDefinition-Patient.json';
 import StructureDefinition from './fixtures/StructureDefinition-StructureDefinition.json';
 import 'fake-indexeddb/auto';
-import { fhirdefs } from 'fsh-sushi';
 
-const FHIRDefinitions = fhirdefs.FHIRDefinitions;
 const defaultConfig = {
   canonical: 'http://example.org',
   version: '1.0.0',
@@ -15,56 +13,48 @@ const defaultConfig = {
 
 describe('#runSUSHI', () => {
   it('should return an undefined package when we get invalid FHIRDefinitions', async () => {
-    const fhirDefs = new FHIRDefinitions();
     const dependencyArr = [];
-    const loadDefsSpy = jest
-      .spyOn(processing, 'loadExternalDependencies')
+    const loadAndCleanDBSpy = jest
+      .spyOn(processing, 'loadAndCleanDatabase')
       .mockReset()
-      .mockResolvedValue({ defs: fhirDefs, emptyDependencies: [] });
-    const checkForDatabaseUpgradeSpy = jest
-      .spyOn(processing, 'checkForDatabaseUpgrade')
-      .mockReset()
-      .mockResolvedValue({ shouldUpdate: false, version: 1 });
+      .mockImplementation((defs, deps) => {
+        // Don't add any FHIR definitions to defs
+        return Promise.resolve(defs);
+      });
     const text =
       'Profile: FishPatient Parent: Patient Id: fish-patient Title: "Fish Patient" Description: "A patient that is a type of fish."';
     const outPackage = await runSUSHI(text, defaultConfig, dependencyArr);
-    expect(loadDefsSpy).toHaveBeenCalled();
-    expect(checkForDatabaseUpgradeSpy).toHaveBeenCalled();
+    expect(loadAndCleanDBSpy).toHaveBeenCalled();
     expect(outPackage).toBeUndefined();
   });
 
   it('should return the correct output package when proper FSH code is entered', async () => {
-    const FHIRDefs = new FHIRDefinitions();
-    FHIRDefs.add(Patient);
-    FHIRDefs.add(StructureDefinition);
-    const loadSpy = jest
-      .spyOn(processing, 'loadExternalDependencies')
+    const loadAndCleanDBSpy = jest
+      .spyOn(processing, 'loadAndCleanDatabase')
       .mockReset()
-      .mockResolvedValue({ defs: FHIRDefs, emptyDependencies: [] });
-    const checkForDatabaseUpgradeSpy = jest
-      .spyOn(processing, 'checkForDatabaseUpgrade')
-      .mockReset()
-      .mockResolvedValue({ shouldUpdate: false, version: 1 });
+      .mockImplementation((defs, deps) => {
+        // Add necessary FHIR definitions to defs
+        defs.add(Patient);
+        defs.add(StructureDefinition);
+        return Promise.resolve(defs);
+      });
     const text =
       'Profile: FishPatient\nParent: Patient\nId: fish-patient\nTitle: "Fish Patient"\n Description: "A patient that is a type of fish."';
     const outPackage = await runSUSHI(text, defaultConfig);
-    expect(loadSpy).toHaveBeenCalled();
-    expect(checkForDatabaseUpgradeSpy).toHaveBeenCalled();
+    expect(loadAndCleanDBSpy).toHaveBeenCalled();
     expect(outPackage.profiles).toHaveLength(1);
   });
 
   it('should return an empty package when fillTank does not execute properly', async () => {
-    const FHIRDefs = new FHIRDefinitions();
-    FHIRDefs.add(Patient);
-    FHIRDefs.add(StructureDefinition);
-    const loadSpy = jest
-      .spyOn(processing, 'loadExternalDependencies')
+    const loadAndCleanDBSpy = jest
+      .spyOn(processing, 'loadAndCleanDatabase')
       .mockReset()
-      .mockResolvedValue({ defs: FHIRDefs, emptyDependencies: [] });
-    const checkForDatabaseUpgradeSpy = jest
-      .spyOn(processing, 'checkForDatabaseUpgrade')
-      .mockReset()
-      .mockResolvedValue({ shouldUpdate: false, version: 1 });
+      .mockImplementation((defs, deps) => {
+        // Add necessary FHIR definitions to defs
+        defs.add(Patient);
+        defs.add(StructureDefinition);
+        return Promise.resolve(defs);
+      });
     const fillTankSpy = jest
       .spyOn(processing, 'fillTank')
       .mockReset()
@@ -73,59 +63,52 @@ describe('#runSUSHI', () => {
       });
     const input = 'Improper FSH code!';
     const outPackage = await runSUSHI(input, defaultConfig);
-    expect(loadSpy).toHaveBeenCalled();
-    expect(checkForDatabaseUpgradeSpy).toHaveBeenCalled();
+    expect(loadAndCleanDBSpy).toHaveBeenCalled();
     expect(fillTankSpy).toHaveBeenCalled();
     expect(outPackage).toBeUndefined();
   });
 });
 
 describe('#runGoFSH', () => {
-  const goFSHDefs = [
-    {
-      resourceType: 'Patient',
-      id: 'MyPatient',
-      name: [
-        {
-          family: 'Smith',
-          given: ['Jane']
-        }
-      ],
-      gender: 'female'
-    }
-  ];
+  const patientDef = {
+    resourceType: 'Patient',
+    id: 'MyPatient',
+    name: [
+      {
+        family: 'Smith',
+        given: ['Jane']
+      }
+    ],
+    gender: 'female'
+  };
+  const goFSHDefs = [JSON.stringify(patientDef)];
 
   it('should return a FSH definition without rules when we load invalid FHIRDefinitions', async () => {
-    const FHIRDefs = new FHIRDefinitions();
     const dependencies = [];
-    const loadDefsSpy = jest
-      .spyOn(processing, 'loadExternalDependencies')
+    const loadAndCleanDBSpy = jest
+      .spyOn(processing, 'loadAndCleanDatabase')
       .mockReset()
-      .mockResolvedValue({ defs: FHIRDefs, emptyDependencies: [] });
-    const checkForDatabaseUpgradeSpy = jest
-      .spyOn(processing, 'checkForDatabaseUpgrade')
-      .mockReset()
-      .mockResolvedValue({ shouldUpdate: false, version: 1 });
+      .mockImplementation((defs, deps) => {
+        // Don't add any FHIR definitions to defs
+        return Promise.resolve(defs);
+      });
     const expectedFSH = ['Instance: MyPatient', 'InstanceOf: Patient', 'Usage: #example'].join('\n');
     const outputFSH = await runGoFSH(goFSHDefs, { dependencies });
-    expect(loadDefsSpy).toHaveBeenCalled();
-    expect(checkForDatabaseUpgradeSpy).toHaveBeenCalled();
+    expect(loadAndCleanDBSpy).toHaveBeenCalled();
     expect(outputFSH).toEqual(expectedFSH);
   });
 
   it('should return a string of FSH when proper JSON is entered and there are valid FHIRDefinitions', async () => {
-    const FHIRDefs = new FHIRDefinitions();
-    FHIRDefs.add(Patient);
-    FHIRDefs.add(StructureDefinition);
     const dependencies = [];
-    const loadSpy = jest
-      .spyOn(processing, 'loadExternalDependencies')
+    const loadAndCleanDBSpy = jest
+      .spyOn(processing, 'loadAndCleanDatabase')
       .mockReset()
-      .mockResolvedValue({ defs: FHIRDefs, emptyDependencies: [] });
-    const checkForDatabaseUpgradeSpy = jest
-      .spyOn(processing, 'checkForDatabaseUpgrade')
-      .mockReset()
-      .mockResolvedValue({ shouldUpdate: false, version: 1 });
+      .mockImplementation((defs, deps) => {
+        // Add necessary FHIR definitions to defs
+        defs.add(Patient);
+        defs.add(StructureDefinition);
+        return Promise.resolve(defs);
+      });
 
     const expectedFSH = [
       'Instance: MyPatient',
@@ -136,10 +119,9 @@ describe('#runGoFSH', () => {
       '* gender = #female'
     ].join('\n');
 
-    const outputFSH = await runGoFSH(goFSHDefs, { dependencies }, FHIRDefs);
+    const outputFSH = await runGoFSH(goFSHDefs, { dependencies });
 
-    expect(loadSpy).toHaveBeenCalled();
-    expect(checkForDatabaseUpgradeSpy).toHaveBeenCalled();
+    expect(loadAndCleanDBSpy).toHaveBeenCalled();
     expect(outputFSH).toEqual(expectedFSH);
   });
 });
